Travis: use requirements.txt, and only install 'test' and 'createcoverage'/
The only other interesting part is code-analysis.
But this can easily fail to build on Travis when you don't have the correct versions for your Python and Plone.
And it shows lots of QA failures locally anyway, so there is no use in having Travis fail because of it.

Well, those are only isort errors, so should be fixable.
But I don't want to put too much in this PR.