Disallow merging automatically in CHANGES.rst.
This can mess things up, putting them under the wrong header when it takes too long to merge a PR.
And CHANGES.rst was missing a header for the 0.8 release anyway, making a wrong merge even more likely.
This is why we use towncrier in Plone, but let's not do that here now.

[ci skip]